requirements.txt
fastapi
uvicorn
moviepy
pydub
elevenlabs
ffmpeg-python
python-decouple
python-multipart
openai
langchain
langchain_openai